Enhance support for local SMTP servers
A common way for sending emails is to use a local SMTP server, most often not requiring
authentication and TLS and providing TLS only with a self-signed certificate (this is, for example,
the default configuration for Postfix on Debian/Ubuntu). Thus, skip authentication if no SMTP
credentials are set and disable TLS if the SMTP address is set to localhost. Along the way remove
the obsolete mail initializer.