Generalise Admin::PostsController#update, Start on change-owner feature
The `update` method was a bit specific, only allowing change of body,
and state (active). Generalised the update method to allow updating of
all whitelisted parameter (through `post_params`).

Added a preliminary modal for allowing change of post ownership.

Added a validation on `Post` to ensure that it belongs to an existing
`User`, previously it would check that a `user_id` was set, but not that
that `user_id` actually relates to an existing user.