Use the Ruby 'pluralize' with locale to show correctly in all languages
What pluralize does depends on the language in use. Pass the locale
to the function so that it can work correctly and pluralizes only words
in languages it knows how to.

This fixes the bug that e.g in Finnish all words had an 's' at the end
and thus became nonsense in Finnish.

Also remove the count '2' – it should only be passed if there is a variable
that actually can change between 1 and 2 or more. The way pluralize is
used here is to always be in plural form.

Note that this pluralize only works for languages that have pluralize
rules defined. By default Ruby only does it for English. Support can be
extended using libraries like
https://github.com/davidcelis/inflections/tree/master/lib/inflections

This commit does not introduce any new inflections. This just makes
sure pluralization is done only in the locale where it works, thus this
change fixes a bug visible in all other languages but English.