test: with theme specified in ENV HELPY_THEME
It is easy to break things in themes - for example with a typo in your views.
Allow running tests with a theme set in the HELPY_THEME environment variable.

Obviously some themes will not implement all functionality
or use different markup so that tests fail due to that.
Thinking about adding a skip_for_themes test helper for that.