Add honeypot to new ticket form
Background:

Helpy gives the option to add a recaptcha to the new ticket form (topics/new). If no re-captcha is configured, we should still add some kind of honeypot anti-spam field to the new ticket form. Obviously this could be hidden if the re-captcha is displayed.


Implementation:

This commit adds a URL input to the new topic form, which will hopefully attract spammers to fill in their website. The form is hidden from regular users using `display: none`. In the controller we're then checking if the input was filled. A regular user wouldn't see the input, i.e. it should remain empty. A spammer in contrast happily adds some nonsense and we can reject the input.

When detecting a spammer we're doing the same thing as when the captcha wasn't filled properly, i.e. rendering the form again.


Potential problems:

- I am not sure, if this "feature" is a problem accessibility wise. I simply don't know if screen readers would also "show" the honeypot or not.
- I've seen this approach work reliably on other websites. However, an attacker, who's directly targetting helpy will of course be able to work around the honey pot quite easily.