Fix 404 after failed submitting new topic with failed captcha
Background:

Anonymous users may enter new topics. Per configuration, this feature is protected by a captcha. When the user fails the captcha test, we want to show the input form again, so that they may re-enter their data and the captcha

Problem:

Currently helpy renders a 404 due to a missing template for `topics/create` when the captcha verification fails. This is happening due to operator precedence in combination with optional parenthesis in ruby. The code

```
render :new && return
```

is interpreted as

```
render(:new && return)
```

I.e. it's returning before `render` is called. Therefore rails defaults to rendering the current action, which is `create` and there's no `create` template. Consequently Rails renders a 404.

Solution:

There are two common solutions to this:

```
render(:new) && return
```

and

```
return render :new
```

Both version ensure, that the render call is executed before the `return`. I've chosen the latter, because this adheres to the preference in the controller of not using parenthesis.

This change also adds a test, that exposes the bug.