Add tags API to create and manage tags
Background:

Admins may manage tags using the admin UI, but they are not managable via API.


This commit adds a tags API to read, create, and update tags for admin users.

I did not implement deleting of tags, since many of the other APIs also don't implement deletion and I thought, that there might be a good reason for that.

The tags API is only available to admins, similarly to the tags admin user interface.

I'm only exposing the basic attributes of tags, that are also available in the user interface.